feat: separate shortcodes section to a component
