+++ packages/hoppscotch-common/src/pages/profile.vue
@@ -158,6 +158,7 @@
                     </form>
                   </div>
                 </section>
+
                 <section class="p-4">
                   <h4 class="font-semibold text-secondaryDark">
                     {{ t("settings.sync") }}
@@ -192,90 +193,8 @@
                     </div>
                   </div>
                 </section>
-                <section class="p-4">
-                  <h4 class="font-semibold text-secondaryDark">
-                    {{ t("settings.short_codes") }}
-                  </h4>
-                  <div class="my-1 text-secondaryLight">
-                    {{ t("settings.short_codes_description") }}
-                  </div>
-                  <div class="relative flex-shrink-0 py-4 overflow-x-auto">
-                    <div
-                      v-if="loading"
-                      class="flex flex-col items-center justify-center"
-                    >
-                      <SmartSpinner class="mb-4" />
-                      <span class="text-secondaryLight">{{
-                        t("state.loading")
-                      }}</span>
-                    </div>
-                    <div
-                      v-if="!loading && myShortcodes.length === 0"
-                      class="flex flex-col items-center justify-center p-4 text-secondaryLight"
-                    >
-                      <img
-                        :src="`/images/states/${colorMode.value}/add_files.svg`"
-                        loading="lazy"
-                        class="inline-flex flex-col object-contain object-center w-16 h-16 mb-8"
-                        :alt="`${t('empty.shortcodes')}`"
-                      />
-                      <span class="mb-4 text-center">
-                        {{ t("empty.shortcodes") }}
-                      </span>
-                    </div>
-                    <div v-else-if="!loading">
-                      <div
-                        class="hidden w-full border-t rounded-t bg-primaryLight lg:flex border-x border-dividerLight"
-                      >
-                        <div class="flex w-full overflow-y-scroll">
-                          <div class="table-box">
-                            {{ t("shortcodes.short_code") }}
-                          </div>
-                          <div class="table-box">
-                            {{ t("shortcodes.method") }}
-                          </div>
-                          <div class="table-box">
-                            {{ t("shortcodes.url") }}
-                          </div>
-                          <div class="table-box">
-                            {{ t("shortcodes.created_on") }}
-                          </div>
-                          <div class="justify-center table-box">
-                            {{ t("shortcodes.actions") }}
-                          </div>
-                        </div>
-                      </div>
-                      <div
-                        class="flex flex-col items-center justify-between w-full overflow-y-scroll border rounded max-h-sm lg:rounded-t-none lg:divide-y border-dividerLight divide-dividerLight"
-                      >
-                        <ProfileShortcode
-                          v-for="(shortcode, shortcodeIndex) in myShortcodes"
-                          :key="`shortcode-${shortcodeIndex}`"
-                          :shortcode="shortcode"
-                          @delete-shortcode="deleteShortcode"
-                        />
-                        <SmartIntersection
-                          v-if="hasMoreShortcodes && myShortcodes.length > 0"
-                          @intersecting="loadMoreShortcodes()"
-                        >
-                          <div
-                            v-if="adapterLoading"
-                            class="flex flex-col items-center py-3"
-                          >
-                            <SmartSpinner />
-                          </div>
-                        </SmartIntersection>
-                      </div>
-                    </div>
-                    <div
-                      v-if="!loading && adapterError"
-                      class="flex flex-col items-center py-4"
-                    >
-                      <component :is="IconHelpCircle" class="mb-4 svg-icons" />
-                      {{ getErrorMessage(adapterError) }}
-                    </div>
-                  </div>
-                </section>
+
+                <ProfileShortcodes />
 
                 <section class="p-4">
                   <h4 class="font-semibold text-secondaryDark text-red-500">
@@ -318,9 +237,6 @@
 
 <script setup lang="ts">
 import { ref, watchEffect, computed } from "vue"
-import { pipe } from "fp-ts/function"
-import * as TE from "fp-ts/TaskEither"
-import { GQLError } from "~/helpers/backend/GQLClient"
 import {
   currentUser$,
   probableUser$,
@@ -339,11 +255,9 @@ import { usePageHead } from "@composables/head"
 
 import { toggleSetting } from "~/newstore/settings"
 import ShortcodeListAdapter from "~/helpers/shortcodes/ShortcodeListAdapter"
-import { deleteShortcode as backendDeleteShortcode } from "~/helpers/backend/mutations/Shortcode"
 
 import IconVerified from "~icons/lucide/verified"
 import IconSettings from "~icons/lucide/settings"
-import IconHelpCircle from "~icons/lucide/help-circle"
 
 type ProfileTabs = "sync" | "teams"
 
@@ -424,14 +338,6 @@ const sendEmailVerification = () => {
 }
 
 const adapter = new ShortcodeListAdapter(true)
-const adapterLoading = useReadonlyStream(adapter.loading$, false)
-const adapterError = useReadonlyStream(adapter.error$, null)
-const myShortcodes = useReadonlyStream(adapter.shortcodes$, [])
-const hasMoreShortcodes = useReadonlyStream(adapter.hasMoreShortcodes$, true)
-
-const loading = computed(
-  () => adapterLoading.value && myShortcodes.value.length === 0
-)
 
 onLoggedIn(() => {
   adapter.initialize()
@@ -443,41 +349,4 @@ onAuthEvent((ev) => {
     return
   }
 })
-
-const deleteShortcode = (codeID: string) => {
-  pipe(
-    backendDeleteShortcode(codeID),
-    TE.match(
-      (err: GQLError<string>) => {
-        toast.error(`${getErrorMessage(err)}`)
-      },
-      () => {
-        toast.success(`${t("shortcodes.deleted")}`)
-      }
-    )
-  )()
-}
-
-const loadMoreShortcodes = () => {
-  adapter.loadMore()
-}
-
-const getErrorMessage = (err: GQLError<string>) => {
-  if (err.type === "network_error") {
-    return t("error.network_error")
-  } else {
-    switch (err.error) {
-      case "shortcode/not_found":
-        return t("shortcodes.not_found")
-      default:
-        return t("error.something_went_wrong")
-    }
-  }
-}
 </script>
-
-<style lang="scss" scoped>
-.table-box {
-  @apply flex flex-1 items-center px-4 py-2 truncate;
-}
-</style>
